backend/Test Execution Microservice/requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==2.1.5
MouseInfo==0.1.3
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
pillow==10.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytweening==1.2.0
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
